Extract portable rendering core out of main.ts (issue #183, phase 1)
Moves the already-Obsidian-free rasterization pipeline (WorkerPool,
ImageConverter, shared worker pool management) into src/render/imageConverter.ts,
and factors SupernoteEmbed's bytes->parsed-note->page-DOM logic into
src/render/noteRenderer.ts (parseNote, buildNotePageElements), built with
plain DOM APIs instead of Obsidian's createDiv/createEl helpers.

Pure refactor with no behavior change: SupernoteEmbed now calls into these
helpers instead of duplicating the logic inline. This is the first step
toward pulling SupernoteView out into a standalone web component, per the
plan posted on #183 - these two modules have zero `obsidian` imports and
could run unmodified in a plain browser.
